chore(earnings): standardize earnings_calendar test structure
Resolves #116

- Move earnings calendar test from non-standard tests/unit/ to
  tests/mcp/test_earnings_calendar.py (matches tests/<pkg>/ convention)
- Route through a real FastMCP instance via _fastmcp_helpers instead of
  the CaptureMCP stub, so production code is exercised under coverage
  (earnings_calendar.py: 71% -> 82%)
- Add cases for missing dates, negative horizon, invalid symbols, and
  DataFrame-like (.loc) calendars; yfinance remains fully mocked
- Drop the orphaned tests/unit/ directory (no __init__.py, single file)
- Ignore .claude/worktrees/ local development workspaces
